server/src/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coverage==5.2.1
cryptography==3.1
distlib==0.3.1
distro==1.5.0
Django==3.1.1
djangorestframework==3.11.1
docker==4.3.1
docker-compose==1.27.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.5.0
numpy==1.19.2
packaging==20.4
paramiko==2.7.2
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-cov==2.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
texttable==1.6.3
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.31
websocket-client==0.57.0
Werkzeug==1.0.1
